bruh

+++ cses/ferris_wheel.cpp
@@ -5,9 +5,9 @@
 using namespace std;
 
 int main() {
-    int n;
+    long long n;
     long long x;
-    scanf("%d %lld", &n, &x);
+    scanf("%lld %lld", &n, &x);
     vector<long long> child;
 
     for (int i = 0; i < n; i++) {
@@ -18,25 +18,25 @@ int main() {
 
     sort(child.begin(), child.end());
 
-    int count = 0;
+    long long count = 0;
 
-    for (int i = child.size() - 1; i >= 0; i--) {
+    for (long long i = child.size() - 1; i >= 0; i--) {
         /* for (long long i : child) { */
         /*     printf("%lld\n", i); */
         /* } */
         /* printf("\n"); */
         long long gondola = x;
-        int c = i;
+        long long c = i;
         if (child[i] == -1) continue;
         gondola = x - child[i];
         /* printf("%lld\n%lld\n%lld\n\n", gondola, child[i], child[c]); */
         child[i] = -1;
     
-        // since child is already sorted, binary search in O(log n) maybe ???
-        for (int j = c; j >= 0 && (gondola != 0); j--) if (child[j] != child[c] && child[j] <= gondola) {
+        // since child is already sorted, binary search in O(log n) would make it faster maybe ???
+        for (long long j = c; j >= 0 && (gondola != 0); j--) if (child[j] != child[c] && child[j] <= gondola) {
             if (gondola - child[j] < 0) continue;
             gondola -= child[j];
-            /* printf("%lld %d -> %d\n", gondola, i, j); */
+            /* printf("%lld k%d -> %d\n", gondola, i, j); */
             child[j] = -1;
             break;
         }
@@ -44,5 +44,5 @@ int main() {
         ++count;
     }
 
-    printf("%d\n", count);
+    printf("%lld\n", count);
 }
